+++ src/ledger/AccountRuleHelperImpl.h
@@ -28,6 +28,9 @@ class AccountRuleHelperImpl : public AccountRuleHelper, NonCopyable
 
     Database& getDatabase() override;
 
+    LedgerDelta*
+    getLedgerDelta() override;
+
     std::vector<AccountRuleFrame::pointer>
     loadAccountRules(std::vector<uint64_t> const ruleIDs) override;